tests: views unit tests to integration tests
